Broke up the create output function
