+++ src/file-handling/models/file-data.ts
@@ -1,0 +1,8 @@
+import { DateCount } from '../../generator/models/date-count';
+
+export interface FileData {
+    [repoName: string]: {
+        stars: DateCount;
+        cumulative: DateCount;
+    };
+}
